requirements.txt
fastapi
uvicorn
crewai
pydantic
crewai[tools]
wikipedia
langchain-qdrant
sentence-transformers
python-dotenv
